AI Clothes Category Website/requirements.txt
flask
gunicorn
tensorflow-cpu
Pillow
numpy